Feat: <Link>의 path에 변수 추가
