+++ src/pages/post/Feeds.jsx
@@ -1,0 +1,9 @@
+import PostButton from './PostButton.jsx';
+
+function Feeds() {
+    return(
+    <PostButton/>
+    )
+}
+
+export default Feeds;
